Write own file for Tooltip component
