+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jsmpc",
-  "version": "0.2.2",
+  "version": "0.1.3",
   "private": true,
   "homepage": "https://github.com/",
   "dependencies": {
